Feat: Publish LTA 135 for November 2024
The show starts with some quick followup on a few recently covered and long-running stories before diving into a big serving of regulatory stories, a handful legal story, just one piece of HR news, and some highlights from Apple's services and original content. The three main stories are a new security feature that appeared unannounced in iOS 18.1, 3rd-party sharing for Find My, and some updates to Apple's Pro apps. The show finishes with a rundown of a few other quick Apple-related stories that made the news in November.