Feat: publish LTA 137
The show for January 2025.

The show starts with some quick followup on a few recently covered and long-running stories before diving into the regulatory and legal stories, a little HR news, and some highlights from Apple's services and original content. The three main stories are President Trump's impact on Apple, the latest OS point-releases, and Apple's Q1 2025 earnings call. The show finishes with a rundown of a few other quick Apple-related stories that made the news in January.