Feat: Publish LTA 139
The show for March 2025

The show starts with some quick followup on a few recently covered and long-running stories before diving into the regulatory and legal stories, a little Apple HR news, and some highlights from Apple's services and original content. The three main stories are President Trump's (not-)reciprocal tariffs, a big shakeup in the Apple Intelligence and Siri teams, and the release of this OS cycle's fourth round of feature updates. The show finishes with a rundown of a few other quick Apple-related stories that made the news in March.