Feat: published LTA 141 for May 2025
This solo show starts with just a few quick followups on long-running stories we're keeping an eye on before exploring just one new regulation and some news Apple indirectly made in a Google court case before moving on to a quick rundown of some stories related to Apple's services and original content. The four main stories are the on-going US tariff omni-shambles, the ongoing regulatory and legal attacks on the App Store's anti-steering rules on both sides of the Atlantic, an impressive set of accessibility enhancements coming later in the year, and the launch of CarPlay Ultra. The show finishes with just a handful of other quick stories that made the news in May.