Feat: published LTP 141 on alt text
In this solo show Bart tries to convince you to do the considerate thing and add alternative text to all the photos you share on social media. Sure, do it because it's the right things to do, but don't think of it as a chore, embrace it as an opportunity to improve your photography!