Feat: publish LTP 142
The show for July 2025 "Lean into Phone Camera Strengths"

This solo show is dedicated to a simple piece of advice — don't try to use a phone camera as if it's a high-end DSLR, it isn't, so let that reality guide your use of this wonderful photographic tool. Avoid the the inherent weaknesses, and lean into the inherent strengths!