Feat: publish LTA 144 for August 2025
August is the so-called 'silly season', and it shows, with just two main stories, but lots more detail than usual in the other regular segments. Starting as usual with feedback and followups the most notable developments are in the secret UK encryption case and the Apple-Epic saga, there are some more regulatory problems for Apple, just one legal story, then the usual HR news before checking in on Apple's services and original content where there is a very notable price change to discuss. The two main stories are Tim Cook's 'Made in America' charm offensive, and Apple's US blood oxygen workaround. As usual the show finish with a few smaller stories that made the month's news.