Feat: Publish LTA 145 for September 2025
The show starts with a quick update on the UK iCloud saga, a handful of regulatory and legal stories, a little HR and acquisition news, and rundown of the notable Apple services and original content news before digging into the three main stories. The first main story is a short look at some Trump-related developments in the US before getting stuck into the the twin big stories, the new hardware from the 'Awe Dropping' event, this year's big new OS releases.The show finishes with the usual rundown of a few other quick Apple-related stories that made the news in September.